main.py - logging -> 삭제 후 따로 매서드 처리 -  테이블 false시 예외처리
setting.json
 - Drop 기능 추가

모든 파일
logging -> logger로 변경

logger.py 추가

AutoCreate.py
Drop 기능 세분화로 len(data)로 drop true/false 여부 판단

query.py
drop기능 query.py에서 판단

+++ Common/Output/DB/Postgresql/AutoCreateTable/Query.py
@@ -1,7 +1,9 @@
-def QueryPlugIn(name):
-    dropT = """DROP TABLE {} ;""".format(name)
-    
-    dropS = """DROP SEQUENCE seq_{}_num;""".format(name)
+def QueryPlugIn(name, all_drop_use, drop_use):
+    if all_drop_use == 'true' :
+        if drop_use =='true' :
+            dropT = """DROP TABLE {} ;""".format(name)
+            
+            dropS = """DROP SEQUENCE seq_{}_num;""".format(name)
     
     createS = """CREATE SEQUENCE seq_{}_num START 1""".format(name)
     
@@ -248,8 +250,12 @@ def QueryPlugIn(name):
                 CONSTRAINT vulnerability_judge_pk PRIMARY KEY (vulnerability_judge_num),
                 CONSTRAINT vulnerability_judge_un UNIQUE (classification_cid)
                 );"""
-    
-                    
-    QUERYLIST = [dropT, dropS, createS, createT]
-    
+    if all_drop_use == 'true' :
+        if drop_use == 'true' :
+            QUERYLIST = [dropT, dropS, createS, createT]
+        elif drop_use == 'false' :
+            QUERYLIST = [createS, createT]
+    elif all_drop_use == 'false' :
+        QUERYLIST = [createS, createT]
+        
     return QUERYLIST
